requirements.txt
elevate==0.1.3
dnspython==2.0.0
